[Core] do nothing
added core notcasting based on BLM's version. moved BLM's version to core and added AST
