+++ locale/de/messages.json
@@ -102,10 +102,6 @@
   "blm.leylines.timestamp-header": "Zeitstempel",
   "blm.leylines.title": "Ley-Linien",
   "blm.leylines.uptime-header": "Uptime",
-  "blm.notcasting.duration-header": "Dauer",
-  "blm.notcasting.timelinelink-button": "Springe zur Zeitleiste",
-  "blm.notcasting.timestamp-header": "Zeitstempel",
-  "blm.notcasting.title": "Zeiten in denen du untätig warst",
   "blm.ogcd-downtime.ogcd-cd-metric": "Aktiviere deine OGCDs immer, sobald sie verfügbar sind, jedoch ohne deine anderen Aktionen zu verzögern. Nutze deine Procs oder schnelles Eisga/Feuga, um OGCDs ohne Verzögerung einzusetzen. Um Zufall auszugleichen werden dir {0} Sekunden gewährt, um deine Abklingzeiten zu verzögern.",
   "blm.procs.suggestions.dropped-f3ps.content": "Du hast mindestens ein <0/> Proc verloren, da du <1/> auslaufen gelassen hast, ohne es zu verwenden.",
   "blm.procs.suggestions.dropped-f3ps.why": "",
@@ -355,6 +351,10 @@
   "core.interrupts.title": "Unterbrochene Anwendungen",
   "core.lucid-dreaming.suggestion.content": "",
   "core.lucid-dreaming.suggestion.why": "",
+  "core.notcasting.duration-header": "",
+  "core.notcasting.timelinelink-button": "",
+  "core.notcasting.timestamp-header": "",
+  "core.notcasting.title": "",
   "core.overheal.direct.name": "Direkt",
   "core.overheal.header.percenttotal": "",
   "core.overheal.header.percenttype": "",
