+++ locale/en/messages.json
@@ -102,10 +102,6 @@
   "blm.leylines.timestamp-header": "Timestamp",
   "blm.leylines.title": "Ley Lines",
   "blm.leylines.uptime-header": "Uptime",
-  "blm.notcasting.duration-header": "Duration",
-  "blm.notcasting.timelinelink-button": "Jump to Timeline",
-  "blm.notcasting.timestamp-header": "Timestamp",
-  "blm.notcasting.title": "Times you did literally nothing",
   "blm.ogcd-downtime.ogcd-cd-metric": "Always make sure to use your OGCDs when they are up but don't clip them. Utilize your procs or fast Blizzard IIIs or Fire IIIs to weave them. To account for random factors you are given a buffer of {0} seconds per instance to hold your cooldowns.",
   "blm.procs.suggestions.dropped-f3ps.content": "You lost at least  one <0/> proc by allowing <1/> to expire without using it.",
   "blm.procs.suggestions.dropped-f3ps.why": "{droppedFireStarters, plural, one {# Firestarter proc} other {# Firestarter procs}} expired.",
@@ -355,6 +351,10 @@
   "core.interrupts.title": "Interrupted Casts",
   "core.lucid-dreaming.suggestion.content": "Try to keep <0/> on cooldown for better MP management.",
   "core.lucid-dreaming.suggestion.why": "{usesMissed, plural, one {# use} other {# uses}} of Lucid Dreaming {usesMissed, plural, one {was} other {were}} missed by holding it for at least a total of {0}.",
+  "core.notcasting.duration-header": "Duration",
+  "core.notcasting.timelinelink-button": "Jump to Timeline",
+  "core.notcasting.timestamp-header": "Timestamp",
+  "core.notcasting.title": "Times you did literally nothing",
   "core.overheal.direct.name": "Direct",
   "core.overheal.header.percenttotal": "% of total overheal",
   "core.overheal.header.percenttype": "Overheal % per type",
