+++ locale/fr/messages.json
@@ -102,10 +102,6 @@
   "blm.leylines.timestamp-header": "Horodatage",
   "blm.leylines.title": "Manalignements",
   "blm.leylines.uptime-header": "Durée de l'effet",
-  "blm.notcasting.duration-header": "Durée",
-  "blm.notcasting.timelinelink-button": "Voir sur la chronologie",
-  "blm.notcasting.timestamp-header": "Horodatage",
-  "blm.notcasting.title": "Temps passé à ne rien faire",
   "blm.ogcd-downtime.ogcd-cd-metric": "Faites en sorte de toujours utiliser vos actions oGCDs lorsqu'elles sont disponibles, mais ne les clippez pas. Utilisez les effets additionnels déclenchés par vos sorts ou profitez des incantations rapides de Méga Glace ou de Méga Feu durant les changements de phase pour les utiliser de façon optimale. Afin de tenir compte l'aléatoire, une fenêtre de {0} secondes par fenêtre d'utilisation de vos cooldowns vous est accordée dans le calcul de ces stats.",
   "blm.procs.suggestions.dropped-f3ps.content": "Vous avez perdu un proc de <0/> car vous avez laissé <1/> expirer sans l'utiliser.",
   "blm.procs.suggestions.dropped-f3ps.why": "{droppedFireStarters, plural, one {# instance} other {# instances}} d'expiration de Pyromane.",
@@ -355,6 +351,10 @@
   "core.interrupts.title": "Incantations interrompues",
   "core.lucid-dreaming.suggestion.content": "Essayer de garder <0/> en rechargement pour une meilleure gestion de votre mana.",
   "core.lucid-dreaming.suggestion.why": "{usesMissed, plural, one {# utilisation} other {# utilisations}} de Rêve lucide {usesMissed, plural, one {a été manquée} other {ont été manquées}} en le laissant disponible pendant au moins {0}.",
+  "core.notcasting.duration-header": "",
+  "core.notcasting.timelinelink-button": "",
+  "core.notcasting.timestamp-header": "",
+  "core.notcasting.title": "",
   "core.overheal.direct.name": "Direct",
   "core.overheal.header.percenttotal": "% de sur-soin",
   "core.overheal.header.percenttype": "% de sur-soin par type",
