+++ locale/ja/messages.json
@@ -102,10 +102,6 @@
   "blm.leylines.timestamp-header": "タイムスタンプ",
   "blm.leylines.title": "黒魔紋",
   "blm.leylines.uptime-header": "アクティブ率",
-  "blm.notcasting.duration-header": "効果時間",
-  "blm.notcasting.timelinelink-button": "タイムラインに移動",
-  "blm.notcasting.timestamp-header": "タイムスタンプ",
-  "blm.notcasting.title": "何もしなかった時間。",
   "blm.ogcd-downtime.ogcd-cd-metric": "アビリティを使えるときは必ず使用することにしましょう。ただし、GCDに食い込まないようにしてください。procもしくは詠唱の早いブリザジャやファイガを利用したときに使いましょう。ランダム要素は分析の範囲内に計算を含めてあるため、アクションやアビリティの抱えられる可能な時間を一回につき{0} です。",
   "blm.procs.suggestions.dropped-f3ps.content": "<1/>を使用せずに時間切れで失ったために、1つ以上の<0/>procを失いました。",
   "blm.procs.suggestions.dropped-f3ps.why": "{droppedFireStarters, plural, other {# 個のファイガproc}}が失われた。",
@@ -355,6 +351,10 @@
   "core.interrupts.title": "詠唱中断",
   "core.lucid-dreaming.suggestion.content": "MP管理を良くするために、<0/>をリキャストごとに使いましょう。",
   "core.lucid-dreaming.suggestion.why": "",
+  "core.notcasting.duration-header": "",
+  "core.notcasting.timelinelink-button": "",
+  "core.notcasting.timestamp-header": "",
+  "core.notcasting.title": "",
   "core.overheal.direct.name": "直接",
   "core.overheal.header.percenttotal": "",
   "core.overheal.header.percenttype": "",
