+++ locale/ko/messages.json
@@ -102,10 +102,6 @@
   "blm.leylines.timestamp-header": "타임 스탬프",
   "blm.leylines.title": "흑마법 문양",
   "blm.leylines.uptime-header": "적용 시간",
-  "blm.notcasting.duration-header": "지속 시간",
-  "blm.notcasting.timelinelink-button": "해당 타임라인으로 이동",
-  "blm.notcasting.timestamp-header": "타임 스탬프",
-  "blm.notcasting.title": "정말 말 그대로 아무것도 하지 않았던 구간",
   "blm.ogcd-downtime.ogcd-cd-metric": "논글쿨 기술이 놀지 않도록 상시로 사용하되, 글쿨이 밀리지 않게끔 주의하십시오. 각종 프록이나 빠른 블리자가, 빠른 파이가를 최대한 활용하십시오. 각종 요소를 감안하여 인스턴스 당 기술을 쓰기 전 보류 시간을 {0}초 부여하고 있습니다.",
   "blm.procs.suggestions.dropped-f3ps.content": "<1/>이 끊기기 전까지 효과 향상된 <0/>을 사용하지 않아 최소 1회 이상 사용 기회를 놓쳤습니다.",
   "blm.procs.suggestions.dropped-f3ps.why": "파이가 효과 향상 프록이 {droppedFireStarters, plural, other {#개}} 만료되었습니다.",
@@ -355,6 +351,10 @@
   "core.interrupts.title": "중단된 시전 기술",
   "core.lucid-dreaming.suggestion.content": "",
   "core.lucid-dreaming.suggestion.why": "",
+  "core.notcasting.duration-header": "",
+  "core.notcasting.timelinelink-button": "",
+  "core.notcasting.timestamp-header": "",
+  "core.notcasting.title": "",
   "core.overheal.direct.name": "직격",
   "core.overheal.header.percenttotal": "",
   "core.overheal.header.percenttype": "",
