+++ locale/zh/messages.json
@@ -102,10 +102,6 @@
   "blm.leylines.timestamp-header": "时间截",
   "blm.leylines.title": "黑魔纹",
   "blm.leylines.uptime-header": "覆盖率",
-  "blm.notcasting.duration-header": "持续时间",
-  "blm.notcasting.timelinelink-button": "跳转到时间轴",
-  "blm.notcasting.timestamp-header": "时间截",
-  "blm.notcasting.title": "你在发呆的时间",
   "blm.ogcd-downtime.ogcd-cd-metric": "确保在你的能力技转好时就使用它们，但是不要卡gcd释放。使用你的触发或者半咏唱的爆炎/冰封来插入这些能力技。在考虑随机因素的影响下，你每场战斗有{0} 的时间来延后你冷却好的能力技，",
   "blm.procs.suggestions.dropped-f3ps.content": "至少有一个<0/>触发的<1/>因未被使用而损失掉。",
   "blm.procs.suggestions.dropped-f3ps.why": "损失了 {droppedFireStarters, plural, one {# 次} other {# 次}}火苗效果。",
@@ -355,6 +351,10 @@
   "core.interrupts.title": "中断的吟唱",
   "core.lucid-dreaming.suggestion.content": "尽可能保持 <0/> 不空转以增加可用的MP。",
   "core.lucid-dreaming.suggestion.why": "整场战斗共计可用 {0} 次醒梦，但由于空转而少用了 {usesMissed, plural, other {#}} {usesMissed, plural, other {次}}。",
+  "core.notcasting.duration-header": "",
+  "core.notcasting.timelinelink-button": "",
+  "core.notcasting.timestamp-header": "",
+  "core.notcasting.title": "",
   "core.overheal.direct.name": "直接治疗",
   "core.overheal.header.percenttotal": "%治疗过量",
   "core.overheal.header.percenttype": "各种过量治疗占比",
