+++ src/parser/core/modules/DISPLAY_ORDER.ts
@@ -5,12 +5,13 @@ enum DISPLAY_ORDER {
 	SUGGESTIONS = -20,
 	STATISTICS = -10,
 	// Module DO is 0-100
-	COMBOS = 110,
-	POSITIONALS = 120,
-	DEFENSIVES = 125,
-	TIMELINE = 130,
-	EVENTS_VIEW = 140,
-	CHANGELOG = 150,
+	NOT_CASTING = 110,
+	COMBOS = 120,
+	POSITIONALS = 130,
+	DEFENSIVES = 140,
+	TIMELINE = 150,
+	EVENTS_VIEW = 160,
+	CHANGELOG = 170,
 }
 
 export default DISPLAY_ORDER
